% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterStabilityAnalysis.R
\name{consensus_and_stability}
\alias{consensus_and_stability}
\title{Cluster Stability Testing/Quantification
Subset and cluster multiple times (store cluster assignments in table)
Then 1) find a way to assign universal names for clusters
and 2) quantify cluster stability for each cluster, and
also 3) quantify often each data point (cell) is assigned to its primary cluster}
\usage{
consensus_and_stability(clust_iter_list = NULL,
  clust_iter_table = NULL)
}
\arguments{
\item{clust_iter_list}{a list of lists.  First level list is just all the iterations.  Second
level (inner) list is a list of each cluster, each one containing the indices of all the
original datapoints assigned to that cluster}

\item{clust_iter_table}{a matrix with each column as separate iteration of cluster assignments
(same length as full or initially sampled dataset) --note that there should be NAs in this
table because we want clustering iterations from samples of the dataset (to see how stable
clusters are)}
}
\description{
Cluster Stability Testing/Quantification
Subset and cluster multiple times (store cluster assignments in table)
Then 1) find a way to assign universal names for clusters
and 2) quantify cluster stability for each cluster, and
also 3) quantify often each data point (cell) is assigned to its primary cluster
}
