% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessPanel.R
\name{get.clustering.annotate}
\alias{get.clustering.annotate}
\title{Get clustering variables in legible format
#' @param panel panel formatted following generation by getMetadata.R and input by read.panel
function}
\usage{
get.clustering.annotate(panel)
}
\description{
Get clustering variables in legible format
#' @param panel panel formatted following generation by getMetadata.R and input by read.panel
function
}
