% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessPanel.R
\name{get.plotting.annotate}
\alias{get.plotting.annotate}
\title{Get panel variables in legible format}
\usage{
get.plotting.annotate(panel)
}
\arguments{
\item{panel}{panel formatted following generation by getMetadata.R and input by read.panel
function}
}
\description{
Get panel variables in legible format
}
