% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInput.R
\name{get.transform.annotate}
\alias{get.transform.annotate}
\title{Get marker names in more legibile format for Concat_Transformed file}
\usage{
get.transform.annotate(panel)
}
\arguments{
\item{panel}{panel formatted following generation by getMetadata.R and input by read.panel
function}
}
\description{
Get marker names in more legibile format for Concat_Transformed file
}
