% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInput.R
\name{get.transform.markers}
\alias{get.transform.markers}
\title{Pull out markers for transformation}
\usage{
get.transform.markers(panel)
}
\arguments{
\item{panel}{panel formatted following generation by getMetadata.R and input by read.panel
function}
}
\description{
Finds markers in panel file marked for either clustering or plotting
}
