% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterStabilityAnalysis.R
\name{iterate_dbscan_clusters}
\alias{iterate_dbscan_clusters}
\title{Cluster Stability Testing/Quantification
Subset and cluster multiple times (store cluster assignments in table)
Then 1) find a way to assign universal names for clusters
and 2) quantify cluster stability for each cluster, and
also 3) quantify often each data point (cell) is assigned to its primary cluster
Get dbscan clusters across multiple iterations}
\usage{
iterate_dbscan_clusters(input_dataset = NULL, subsample_size = NULL,
  num_iterations = 50, minPts = 50)
}
\arguments{
\item{input_dataset}{matrix to be clustered on}

\item{subsample_size}{number of subsamples to take in each clustering iteration}

\item{num_iterations}{number of rounds of clustering to test before stability analysis}

\item{minPts}{hdbscan parameter}
}
\description{
Cluster Stability Testing/Quantification
Subset and cluster multiple times (store cluster assignments in table)
Then 1) find a way to assign universal names for clusters
and 2) quantify cluster stability for each cluster, and
also 3) quantify often each data point (cell) is assigned to its primary cluster
Get dbscan clusters across multiple iterations
}
