% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterStabilityAnalysis.R
\name{iterate_louvain_clusters}
\alias{iterate_louvain_clusters}
\title{Iterate through multiple rounds of louvain clustering}
\usage{
iterate_louvain_clusters(input_dataset = NULL, subsample_size = NULL,
  num_iterations = 50, graph_knn = 5)
}
\arguments{
\item{input_dataset}{matrix to be clustered on}

\item{subsample_size}{number of subsamples to take in each clustering iteration}

\item{num_iterations}{number of rounds of clustering to test before stability analysis}

\item{graph_knn}{number of neighbors on knn graph for clustering}
}
\description{
Iterate through multiple rounds of louvain clustering
}
