% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInput.R
\name{read.clustering.rep}
\alias{read.clustering.rep}
\title{Clustering replicate file input}
\usage{
read.clustering.rep(input.folder,
  clustering.rep.filename = "ClusterStabilityCheck.csv")
}
\arguments{
\item{input.folder}{directory containing layout file}

\item{clustering.rep.filename}{clustering replicate filename, defaults to
ClusterStabilityCheck.csv as outputted by generation script}
}
\description{
Reads clustering replicates for consensus clustering and stability analysis
}
