% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInput.R
\name{read.clusters}
\alias{read.clusters}
\title{Cluster file input}
\usage{
read.clusters(input.folder, clusters.filename = "clusters.csv")
}
\arguments{
\item{input.folder}{directory containing layout file}

\item{cluster.filename}{clustering replicate filename, defaults to clusters.csv as outputted by
generation script}
}
\description{
Reads clusters from pipeline
}
