% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInput.R
\name{read.concat.transformed}
\alias{read.concat.transformed}
\title{Concat_Transformed file input}
\usage{
read.concat.transformed(input.folder,
  concat.transformed.filename = "Concat_Tranformed.csv")
}
\arguments{
\item{input.folder}{directory containing metadata file}

\item{concat.transformed.filename}{concat_transformed filename, defaults to
Concat_Transformed.csv as outputted by generation script}
}
\description{
Reads Concat_Transformed file for use in Zunder lab pipeline
}
