% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInput.R
\name{read.layout}
\alias{read.layout}
\title{Layout file input}
\usage{
read.layout(input.folder, layout.filename = "UMAP_layout.csv")
}
\arguments{
\item{input.folder}{directory containing layout file}

\item{layout.filename}{layout filename, defaults to UMAP_layout.csv as outputted by generation
script}
}
\description{
Reads uMAP or other 2D layout for plotting
}
