% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInput.R
\name{read.metadata}
\alias{read.metadata}
\title{Read metadata}
\usage{
read.metadata(input.folder, md.filename = "metadata.csv")
}
\arguments{
\item{input.folder}{directory containing metadata file}

\item{md.filename}{metadata filename, defaults to metadata.csv as outputted by generation script}
}
\description{
Reads metadata file and produces format friendly to scripts in Zunder lab pipeline. Metadata
strategy inspired by Nowicka et al., 2017, F1000 Research
}
