% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInput.R
\name{read.panel}
\alias{read.panel}
\title{Panel file input}
\usage{
read.panel(input.folder, panel.filename = "panel.csv")
}
\arguments{
\item{input.folder}{directory containing metadata file}

\item{panel.filename}{panel filename, defaults to panel.csv as outputted by generation script}
}
\description{
Reads panel file and produces format friendly to scripts in Zunder lab pipeline
}
